Refactor to have shared ScriptServiceV3AlphaExecutor base class
