+++ source/Octopus.Tentacle/Util/ScriptLogExtensions.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using Octopus.Tentacle.Contracts;
+using Octopus.Tentacle.Scripts;
 
 namespace Octopus.Tentacle.Util
 {
@@ -16,6 +18,13 @@ namespace Octopus.Tentacle.Util
             log("##octopus[stdout-verbose]");
             log(message);
             log("##octopus[stdout-default]");
+        }
+
+        public static void WriteVerbose(this IScriptLogWriter writer, string message)
+        {
+            writer.WriteOutput(ProcessOutputSource.StdOut, "##octopus[stdout-verbose]");
+            writer.WriteOutput(ProcessOutputSource.StdOut, message);
+            writer.WriteOutput(ProcessOutputSource.StdOut, "##octopus[stdout-default]");
         }
     }
 }
